Move Node.js-based tests to `node-tests` folder
