+++ .npmignore
@@ -1,6 +1,7 @@
 /bower_components
 /config/ember-try.js
 /dist
+/node-tests
 /tests
 /tmp
 **/.gitkeep
