+++ node-tests/assert/file-contains.js
