+++ node-tests/assert/file-exists.js
