+++ node-tests/assert/file-has-content.js
