+++ node-tests/fixtures/default-options.js
