+++ node-tests/fixtures/paths.js
