+++ node-tests/get-directory-contents.js
